Seprated the navbar to diff component
