+++ src/components/Navbar.jsx
@@ -1,13 +1,44 @@
-import React from "react";
+import React, { useState } from "react";
 import { NavLink } from "react-router-dom";
+import filterDuplicate from "../components/utils/Filterduplicate";
 
-const Navbar = () => {
+const Navbar = ({ recipes, searchBar, setSearchBar }) => {
+  const [value, setValue] = useState("");
+
+  const filteredItems = filterDuplicate({ recipes });
+  const handleSearchBar = (e) => {
+    e.preventDefault();
+    setSearchBar(e.target.value);
+    setValue("");
+  };
   return (
-    <div className="navbar">
-      <NavLink to="/">Home</NavLink>
-      <NavLink to="/meat">Meat</NavLink>
-      <NavLink to="/fish">Fish</NavLink>
-      <NavLink to="/vegan">Vegan</NavLink>
+    <div className="container-navbar-searchbar">
+      <div className="container-navbar">
+        <NavLink to="/">Home</NavLink>
+        {filteredItems &&
+          filteredItems.map((item) => (
+            <div key={item.fields.category}>
+              <NavLink to={`/${item.fields.category}`}>
+                {item.fields.category}
+              </NavLink>
+            </div>
+          ))}
+      </div>
+      <div className="container-searchbar-btn">
+        <form>
+          <input
+            id="searchFiled"
+            placeholder="Search by category"
+            onChange={(e) => {
+              setValue(e.target.value);
+            }}
+            value={value}
+          />
+          <button id="searchBtn" onClick={handleSearchBar} value={value}>
+            Search
+          </button>
+        </form>
+      </div>
     </div>
   );
 };
